Made the necessary to the exception and module files after the initial review.
